pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xrobot"
version = "0.2.0"
description = "A lightweight application framework for robot module management, lifecycle control, and hardware abstraction."
requires-python = ">=3.8"
authors = [
    {name = "Xiao", email = "2592509183@qq.com"}
]
maintainers = [
    {name = "Xiao", email = "2592509183@qq.com"}
]
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["code generation", "embedded systems", "automation"]
dependencies = [
    "argparse",
    "pyyaml"
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
cli = ["rich", "click"]

[project.urls]
Homepage = "https://github.com/xrobot-org"
Documentation = "https://xrobot-org.github.io"
Repository = "https://github.com/xrobot-org/XRobot"
"Bug Tracker" = "https://github.com/xrobot-org/XRobot/issues"

[project.scripts]
xrobot_gen_main = "xrobot.GenerateMain:main"
xrobot_mod_parser = "xrobot.ModuleParser:main"
xrobot_create_mod = "xrobot.ModuleCreator:main"
xrobot_init_mod = "xrobot.InitModule:main"
xrobot_setup = "xrobot.XRobotSetup:main"
xrobot_add_mod = "xrobot.AddModule:main"
xrobot_src_man = "xrobot.SourceManager:main"

[tool.setuptools.packages.find]
where = ["src"]
